requirements.txt

tensorflow-io-gcs-filesystem==0.34.0
pillow==10.4.0
numpy==1.26.4

# tensorflow>=2.15.0
# tensorflow-io-gcs-filesystem>=0.34.0
# kfp>=2.1.0,<3
# google-cloud-aiplatform>=1.49.0
# google-cloud-pipeline-components>=2.6.0,<3
# google-cloud-storage>=2.10.0
# pillow>=10.0.0
# numpy>=1.26.0
# cloud-tpu-client>=0.10.0
# typing-extensions>=4.7.1
